Eddy diffusion consolidation; cleanup; documentation
